+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/async/BackgroundTaskQueueTests.cs
@@ -5,7 +5,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace com.csutil.tests.async {
+namespace com.csutil.integrationTests.async {
 
     [Collection("Sequential")] // Will execute tests in here sequentially
     public class BackgroundTaskQueueTests {
