+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/async/TaskDebounceTests.cs
@@ -4,7 +4,7 @@ using System.Diagnostics;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace com.csutil.tests.async {
+namespace com.csutil.integrationTests.async {
 
     [Collection("Sequential")] // Will execute tests in here sequentially
     public class TaskDebounceTests {
