+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/http/HasInternetTests.cs
@@ -2,7 +2,7 @@
 using com.csutil.http;
 using Xunit;
 
-namespace com.csutil.tests.http {
+namespace com.csutil.integrationTests.http {
 
     [Collection("Sequential")] // Will execute tests in here sequentially
     public class HasInternetTests : IHasInternetListener {
