+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/http/XmlParsingTests.cs
@@ -5,7 +5,7 @@ using com.csutil.http;
 using com.csutil.http.apis;
 using Xunit;
 
-namespace com.csutil.tests.http {
+namespace com.csutil.integrationTests.http {
 
     public class XmlParsingTests {
