+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/http/apis/GoogleSheetsV4Tests.cs
@@ -2,7 +2,7 @@
 using com.csutil.http.apis;
 using Xunit;
 
-namespace com.csutil.tests.http {
+namespace com.csutil.integrationTests.http {
 
     public class GoogleSheetsV4Tests {
