+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/http/apis/OpenAiTests.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 using com.csutil.http.apis;
 using Xunit;
 
-namespace com.csutil.tests.http {
+namespace com.csutil.integrationTests.http {
 
     public class OpenAiTests {
