+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/http/apis/WikipediaApiTests.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using com.csutil.http.apis.wikipedia;
 using Xunit;
 
-namespace com.csutil.tests.http {
+namespace com.csutil.integrationTests.http {
 
     public class WikipediaApiTests {
