+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/model/ModelPersistenceTests.cs
@@ -6,7 +6,7 @@ using com.csutil.model;
 using Xunit;
 using Zio;
 
-namespace com.csutil.tests.model {
+namespace com.csutil.integrationTests.model {
 
     [Collection("Sequential")] // Will execute tests in here sequentially
     public class ModelPersistenceTests {
