+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/system/NewsManagerTests.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace com.csutil.tests.system {
+namespace com.csutil.integrationTests.system {
 
     [Collection("Sequential")] // Will execute tests in here sequentially
     public class NewsManagerTests {
